src-ava/tools/gdbTraceParser/PinTraceUtilTest.java
/*******************************************************************************
 *    Copyright 2019 Fabrizio Pastore, Leonardo Mariani
 *   
 *    Licensed under the Apache License, Version 2.0 (the "License");
 *    you may not use this file except in compliance with the License.
 *    You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 *    Unless required by applicable law or agreed to in writing, software
 *    distributed under the License is distributed on an "AS IS" BASIS,
 *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *    See the License for the specific language governing permissions and
 *    limitations under the License.
 *******************************************************************************/
package tools.gdbTraceParser;

import static org.junit.Assert.*;

import java.io.File;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;

public class PinTraceUtilTest {

	@Before
	public void setUp() throws Exception {
	}

	@After
	public void tearDown() throws Exception {
	}

	@Test
	public void testGetPid() {
		File traceFile = new File( "AnalysisV0/BCT/BCT_DATA/BCT/validTraces/bdciTrace.1455541322.28732.0.pin" );
		
		String tid = PinTraceUtil.geThreadId(traceFile);
		
		assertEquals("0", tid);
		
		String pid = PinTraceUtil.getPID(traceFile);
		assertEquals("145554132228732", pid);
	}

}
